refactor(lib): split utils.ts into specialized modules
